requirements.txt
flask
pymongo
python-dotenv
langchain-core
langchain-community
faiss-cpu
sentence-transformers
pypdf2
requests
pandas
openpyxl
google-genai>=1.0.0
